feat: update finalize view and payment / signature processing
