+++ API/CCW.Common/Models/Application.cs
@@ -36,7 +36,7 @@ public class Application
     public WorkInformation WorkInformation { get; set; }
     public QualifyingQuestions QualifyingQuestions { get; set; }
     public int CurrentStep { get; set; }
-    public ApplicationStatus OriginalStatus { get; set; }
+    public ApplicationStatus? OriginalStatus { get; set; }
     public ApplicationStatus Status { get; set; }
     public AppointmentStatus AppointmentStatus { get; set; }
     public string AppointmentId { get; set; }
