+++ API/CCW.Common/Models/BackgroundCheck.cs
@@ -4,153 +4,30 @@ namespace CCW.Common.Models;
 
 public class BackgroundCheck
 {
-    public ProofOfID ProofOfID { get; set; }
-    public ProofOfResidency ProofOfResidency { get; set; }
-    public NCICWantsWarrants NCICWantsWarrants { get; set; }
-    public Locals Locals { get; set; }
-    public Probations Probations { get; set; }
-    public DMVRecord DMVRecord { get; set; }
-    public AKAsChecked AKAsChecked { get; set; }
-    public CrimeTracer CrimeTracer { get; set; }
-    public TrafficCourtPortal TrafficCourtPortal { get; set; }
-    public DOJApprovalLetter DOJApprovalLetter { get; set; }
-    public CIINumber CIINumber { get; set; }
-    public DOJ DOJ { get; set; }
-    public FBI FBI { get; set; }
-    public Livescan Livescan { get; set; }
-    public SR14 SR14 { get; set; }
-    public Firearms Firearms { get; set; }
-    public SidLettersReceived SidLettersReceived { get; set; }
-    public SafetyCertificate SafetyCertificate { get; set; }
-    public Restrictions Restrictions { get; set; }
+    public BackgroundCheckItem ProofOfID { get; set; }
+    public BackgroundCheckItem ProofOfResidency { get; set; }
+    public BackgroundCheckItem NCICWantsWarrants { get; set; }
+    public BackgroundCheckItem Locals { get; set; }
+    public BackgroundCheckItem Probations { get; set; }
+    public BackgroundCheckItem DMVRecord { get; set; }
+    public BackgroundCheckItem AKAsChecked { get; set; }
+    public BackgroundCheckItem CrimeTracer { get; set; }
+    public BackgroundCheckItem TrafficCourtPortal { get; set; }
+    public BackgroundCheckItem DOJApprovalLetter { get; set; }
+    public BackgroundCheckItem CIINumber { get; set; }
+    public BackgroundCheckItem DOJ { get; set; }
+    public BackgroundCheckItem FBI { get; set; }
+    public BackgroundCheckItem Livescan { get; set; }
+    public BackgroundCheckItem SR14 { get; set; }
+    public BackgroundCheckItem Firearms { get; set; }
+    public BackgroundCheckItem SidLettersReceived { get; set; }
+    public BackgroundCheckItem SafetyCertificate { get; set; }
+    public BackgroundCheckItem Restrictions { get; set; }
 }
 
-public class ProofOfID
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class ProofOfResidency
+public class BackgroundCheckItem
 {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class NCICWantsWarrants
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class Locals
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class Probations
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class DMVRecord
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class AKAsChecked
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class CrimeTracer
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class TrafficCourtPortal
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-    public class DOJApprovalLetter
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class CIINumber
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class DOJ
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class FBI
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-    public class Livescan
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-    public class SR14
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class Firearms
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class SidLettersReceived
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class SafetyCertificate
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
-
-    public class Restrictions
-    {
-        public bool? Value { get; set; }
-        public DateTime? ChangeDateTimeUtc { get; set; }
-        public string ChangeMadeBy { get; set; }
-    }
+    public bool? Value { get; set; }
+    public DateTime? ChangeDateTimeUtc { get; set; }
+    public string ChangeMadeBy { get; set; }
+}
