+++ API/CCW.Common/Models/UserApplication.cs
@@ -37,7 +37,7 @@ public class UserApplication
     public WorkInformation WorkInformation { get; set; }
     public QualifyingQuestions QualifyingQuestions { get; set; }
     public int CurrentStep { get; set; }
-    public ApplicationStatus OriginalStatus { get; set; }
+    public ApplicationStatus? OriginalStatus { get; set; }
     public ApplicationStatus Status { get; set; }
     public AppointmentStatus AppointmentStatus { get; set; }
     public string AppointmentId { get; set; }
