+++ UI/libs/core-public/src/lib/components/containers/ModifyFormContainer.vue
@@ -66,20 +66,6 @@
             </div>
           </v-stepper-step>
 
-          <v-divider></v-divider>
-
-          <v-stepper-step
-            editable
-            :complete="stepFiveValid"
-            :edit-icon="stepFiveValid ? 'mdi-check' : '$edit'"
-            :color="stepFiveValid ? 'success' : 'primary'"
-            :step="5"
-          >
-            <div class="text-center">
-              {{ $t('Finalize') }}
-            </div>
-          </v-stepper-step>
-
           <v-progress-linear
             :active="
               isGetApplicationLoading ||
@@ -140,17 +126,6 @@
               @update-step-four-valid="handleUpdateStepFourValid"
               @handle-continue="handleContinueFile"
               @handle-save="handleSaveFile"
-            />
-          </v-stepper-content>
-        </v-stepper-items>
-
-        <v-stepper-items>
-          <v-stepper-content :step="5">
-            <ModifyFinalizeStep
-              :modifying-name="modifyingName"
-              :modifying-address="modifyingAddress"
-              :modifying-weapons="modifyingWeapons"
-              :payment-complete="isModificationPaymentComplete"
             />
           </v-stepper-content>
         </v-stepper-items>
@@ -243,21 +218,6 @@
             />
           </v-expansion-panel-content>
         </v-expansion-panel>
-
-        <v-expansion-panel>
-          <v-expansion-panel-header @click.native="stepIndex.step = 5">
-            {{ $t('Finalize') }}
-          </v-expansion-panel-header>
-
-          <v-expansion-panel-content eager>
-            <ModifyFinalizeStep
-              :modifying-name="modifyingName"
-              :modifying-address="modifyingAddress"
-              :modifying-weapons="modifyingWeapons"
-              :payment-complete="isModificationPaymentComplete"
-            />
-          </v-expansion-panel-content>
-        </v-expansion-panel>
       </v-expansion-panels>
     </v-container>
   </div>
@@ -265,21 +225,18 @@
 
 <script lang="ts" setup>
 import ModifyAddressStep from '@core-public/components/form-stepper/modify-form-steps/ModifyAddressStep.vue'
-import ModifyFinalizeStep from '@core-public/components/form-stepper/modify-form-steps/ModifyFinalizeStep.vue'
 import ModifyNameStep from '@core-public/components/form-stepper/modify-form-steps/ModifyNameStep.vue'
 import ModifySupportingDocumentsStep from '@core-public/components/form-stepper/modify-form-steps/ModifySupportingDocumentsStep.vue'
 import ModifyWeaponStep from '@core-public/components/form-stepper/modify-form-steps/ModifyWeaponStep.vue'
 import { useCompleteApplicationStore } from '@shared-ui/stores/completeApplication'
-import { usePaymentStore } from '@shared-ui/stores/paymentStore'
+import { useRouter } from 'vue-router/composables'
 import {
   CompleteApplication,
   WeaponInfoType,
 } from '@shared-utils/types/defaultTypes'
 import { computed, onMounted, reactive, ref, watch } from 'vue'
 import { useMutation, useQuery } from '@tanstack/vue-query'
-import { useRoute, useRouter } from 'vue-router/composables'
 
-const paymentStore = usePaymentStore()
 const applicationStore = useCompleteApplicationStore()
 const isApplicationValid = ref(false)
 const stepIndex = reactive({
@@ -300,9 +257,7 @@ const modifyingWeapons = computed(() => {
 const stepTwoValid = ref(false)
 const stepThreeValid = ref(false)
 const stepFourValid = ref(false)
-const stepFiveValid = ref(false)
 const router = useRouter()
-const route = useRoute()
 
 const expansionStep = computed({
   get() {
@@ -313,53 +268,7 @@ const expansionStep = computed({
   },
 })
 
-const {
-  mutate: updatePaymentHistory,
-  isLoading: isUpdatePaymentHistoryLoading,
-} = useMutation({
-  mutationFn: ({
-    transactionId,
-    successful,
-    amount,
-    paymentType,
-    transactionDateTime,
-    hmac,
-    applicationId,
-  }: {
-    transactionId: string
-    successful: boolean
-    amount: number
-    paymentType: string
-    transactionDateTime: string
-    hmac: string
-    applicationId: string
-  }) => {
-    return paymentStore.updatePaymentHistory(
-      transactionId,
-      successful,
-      amount,
-      paymentType,
-      transactionDateTime,
-      hmac,
-      applicationId
-    )
-  },
-  onSuccess: () =>
-    applicationStore
-      .getCompleteApplicationFromApi(
-        applicationStore.completeApplication.id,
-        Boolean(route.query.isComplete)
-      )
-      .then(res => {
-        applicationStore.setCompleteApplication(res)
-      }),
-})
-
-const {
-  isLoading: isGetApplicationLoading,
-  isRefetching,
-  refetch: getAllUserApplications,
-} = useQuery(
+const { isLoading: isGetApplicationLoading, isRefetching } = useQuery(
   ['getApplicationsByUser'],
   async () => await applicationStore.getAllUserApplicationsApi(),
   {
@@ -380,9 +289,6 @@ const { isLoading: isUpdateApplicationLoading, mutate: updateMutation } =
 
       return applicationStore.updateApplication()
     },
-    onSuccess: () => {
-      getAllUserApplications()
-    },
   })
 
 const { isLoading: isSaveLoading, mutate: saveMutation } = useMutation({
@@ -398,54 +304,7 @@ onMounted(() => {
   window.scrollTo(0, 0)
   isApplicationValid.value = Boolean(applicationStore.completeApplication.id)
 
-  window.console.log(isApplicationValid.value)
-
   stepIndex.step = applicationStore.completeApplication.application.currentStep
-
-  const transactionId = route.query.transactionId
-  const successful = route.query.successful
-  const amount = route.query.amount
-  const hmac = route.query.hmac
-  const paymentType = route.query.paymentType
-  const applicationId = route.query.applicationId
-  let transactionDateTime = route.query.transactionDateTime
-
-  if (typeof transactionDateTime === 'string') {
-    transactionDateTime = transactionDateTime.replace(':', '%3A')
-    transactionDateTime = transactionDateTime.replace(':', '%3A')
-  }
-
-  if (
-    typeof transactionId === 'string' &&
-    typeof successful === 'string' &&
-    typeof amount === 'string' &&
-    typeof paymentType === 'string' &&
-    typeof transactionDateTime === 'string' &&
-    typeof hmac === 'string' &&
-    typeof applicationId === 'string'
-  ) {
-    updatePaymentHistory({
-      transactionId,
-      successful: Boolean(successful),
-      amount: Number(amount),
-      paymentType,
-      transactionDateTime,
-      hmac,
-      applicationId,
-    })
-  }
-})
-
-const isModificationPaymentComplete = computed(() => {
-  return applicationStore.completeApplication.paymentHistory.some(ph => {
-    return ph.paymentType === 4 && ph.successful === true
-  })
-})
-
-const wasModificationPaymentUnsuccessful = computed(() => {
-  return applicationStore.completeApplication.paymentHistory.some(ph => {
-    return ph.successful === false && ph.paymentType === 4
-  })
 })
 
 function handleSaveName(name) {
@@ -518,8 +377,7 @@ function handleContinueFile() {
 
   updateMutation()
 
-  stepIndex.previousStep = stepIndex.step
-  stepIndex.step += 1
+  router.push('/ModifyFinalize')
 }
 
 function handleAddWeapon(weapon: WeaponInfoType) {
