+++ UI/libs/core-public/src/lib/router/index.ts
@@ -3,6 +3,7 @@ import ApplicationView from '@core-public/views/ApplicationView.vue'
 import FinalizeView from '@core-public/views/FinalizeView.vue'
 import FormView from '@core-public/views/FormView.vue'
 import HomeView from '@core-public/views/HomeView.vue'
+import ModifyFinalizeView from '@core-public/views/ModifyFinalizeView.vue'
 import ModifyFormView from '@core-public/views/ModifyFormView.vue'
 import MoreInformationView from '@core-public/views/MoreInformationView.vue'
 import PenalView from '@core-public/views/PenalView.vue'
@@ -71,6 +72,11 @@ const routes: Array<RouteConfig> = [
     path: Routes.MODIFY_FORM_PATH,
     name: 'ModifyForm',
     component: ModifyFormView,
+  },
+  {
+    path: Routes.MODIFY_FINALIZE_PATH,
+    name: 'ModifyFinalize',
+    component: ModifyFinalizeView,
   },
   {
     // keep this at the very end
